GEOM.CLIP / GEOM.SETUP / GEOM.BINNER: the three contracts, and the gates
Contracts filled to the TEXTURE.TMU standard: laws FOUND separated from laws
CHOSEN, with the citation for each found law and the argument plus the
rejected alternative for each chosen one.

FOUND: the near plane is a whole-primitive rejection (so GEOM.CLIP generates
no vertices at all); the guard band is an upstream saturation in
to_screen_xy, not a clip plane; the scissor test is raster_tri's own early
return, reached by CALLING the scan_bbox extracted from it; the edge function
is affine and the third constant is free by the barycentric identity; the
tile is 16 px; submission order inside a tile is painter order, so the list
is FIFO; and the drain port's units are RASTER.EDGEWALK's -- pixels.

CHOSEN: the backface MODE with default NONE (the winding is NOT ratified
here); the counter split and the reject order; the tile grid anchor;
row-major enumeration; the affine corner trivial-reject; and the overflow
wall.

Recorded as shortfalls rather than solved: GEOM.SETUP does NOT build the
attribute gradients -- rast.cpp's model re-derives each row start with its
own rounding, so a plane-setup block would not be bit-exact, and
RASTER.FRAGMENT interpolates nothing, so there is no consumer. GEOM.BINNER
misses the ledger's "1 bin reference per clock" by 2.83x, MEASURED (561
cycles for 198 references), and the composition inherits
zhao_raster_tile_pipe's one clear per job.

formal_runs.yml records the arena proof green (prove + cover, mutation-
checked). Two ledger V20 sites and two cppcheck findings fixed: every "by
construction" claim now names its enforcer, and the composition driver takes
one packet vector instead of two parallel ones.

+++ tests/geometry/geom_clip_random.cpp
@@ -52,7 +52,8 @@ const Clip::Viewport kVps[4] = {
 std::vector<uint8_t> serialize(const Clip::In& t, const Clip::Viewport& vp, int cull) {
   std::vector<uint8_t> v;
   auto put = [&v](int32_t x) {
-    for (int i = 0; i < 4; ++i) v.push_back(static_cast<uint8_t>(static_cast<uint32_t>(x) >> (8 * i)));
+    for (int i = 0; i < 4; ++i)
+      v.push_back(static_cast<uint8_t>(static_cast<uint32_t>(x) >> (8 * i)));
   };
   put(t.ax);
   put(t.ay);
@@ -74,8 +75,8 @@ std::string describe(const Clip::Out& want, const Clip::Out& got) {
   std::snprintf(buf, sizeof(buf),
                 "verdict oracle %d rtl %d; box oracle [%d..%d]x[%d..%d] rtl [%d..%d]x[%d..%d]; "
                 "2A oracle %lld rtl %lld",
-                want.verdict, got.verdict, want.min_x, want.max_x, want.min_y, want.max_y,
-                got.min_x, got.max_x, got.min_y, got.max_y,
+                static_cast<int>(want.verdict), static_cast<int>(got.verdict), want.min_x,
+                want.max_x, want.min_y, want.max_y, got.min_x, got.max_x, got.min_y, got.max_y,
                 static_cast<long long>(want.area2), static_cast<long long>(got.area2));
   return std::string(buf);
 }
@@ -222,8 +223,8 @@ void lane_b(ClipDev& dev, uint32_t iters) {
       const zref::EdgeWalk::Tri et{t.ax, t.ay, t.bx, t.by, t.cx, t.cy};
       for (uint32_t ty = vp.y0; ty < vp.y0 + vp.h && total == 0; ty += 16)
         for (uint32_t tx = vp.x0; tx < vp.x0 + vp.w && total == 0; tx += 16)
-          total += zref::EdgeWalk::tile(et, static_cast<int32_t>(tx), static_cast<int32_t>(ty))
-                       .count;
+          total +=
+              zref::EdgeWalk::tile(et, static_cast<int32_t>(tx), static_cast<int32_t>(ty)).count;
       if (total != 0) {
         bad = true;
         err = "kOffscreen deleted real coverage";
